+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Forge
+
+#### Added
+
+- `Secp256k1` and `Secp256r1` support for `KeyPair` in `snforge_std`
+
+#### Changed
+
+- `snforge_std::signature` module with `key_pair`, `stark_curve` and `secp_256_curve` submodules
+
 ## [0.12.0] - 2023-12-06
 
 ### Forge